+++ lin_sys/kkt_common/kkt.h
@@ -6,8 +6,6 @@ extern "C" {
 # endif // ifdef __cplusplus
 
 # include "osqp.h"
-# include "types.h"
-# include "lin_alg.h"
 
 # ifndef EMBEDDED
